+++ stimela/cargo/cab/casa_gencal/src/run.py
@@ -1,9 +1,8 @@
 import os
 import sys
 import drivecasa
-casa = drivecasa.Casapy(log2term=True, echo_to_stdout=True, timeout=24*3600) 
-# I set timeout to a day. Not my business to
-# decide how long the process should tak
+import logging
+casa = drivecasa.Casapy(log2term=True, echo_to_stdout=True, timeout=24*3600*10)
 
 sys.path.append("/utils")
 import utils
@@ -27,7 +26,13 @@ for param in cab['parameters']:
 
 script = ['{0}(**{1})'.format(cab['binary'], args)]
 
-_, err = casa.run_script(script, raise_on_severe=False)
+def log2term(result):
+    if result[1]:
+        err = '\n'.join(result[1] if result[1] else [''])
+        failed = err.lower().find('an error occurred running task')>=0
+        if failed:
+            raise RuntimeError('CASA Task failed. See error message above')
+        sys.stdout.write('WARNING:: SEVERE messages from CASA run')
 
-if len(err)>0:
-    raise RuntimeError("Caught severe exception while running CASA task {0}. The error message is above".format(cab['binary']))
+result = casa.run_script(script, raise_on_severe=False)
+log2term(result)