+++ stimela/cargo/cab/cleanmask/src/run.py
@@ -16,7 +16,7 @@ for param in params:
     if param['value'] in [False, None]:
         continue
     elif param['value'] is True:
-        arg = "{0}{1}".format(param["prefix"], param["name"])
+        arg = "{0}{1}".format(cab["prefix"], param["name"])
     else:
         arg = "{0}{1} {2}".format(cab["prefix"], param["name"], param["value"])