+++ stimela/cargo/cab/specfit/src/run.py
@@ -1,6 +1,7 @@
 import os
 import sys
 import re
+import pyfits
 from pyrap.tables import table
 
 sys.path.append("/utils")
@@ -42,5 +43,13 @@ else:
 if lsmname and outlsm:
     sys.stdout.write("Extracting spi from image")
 
+    with pyfits.open(image) as hdu:
+        header = hdu[0].header
+        bmin = header['BMIN']
+        bmaj = header['BMAJ']
+        bpa = header['BPA']
+    beam = (bmin, bmaj, bpa)
+
     import addSPI
-    addSPI.addSPI(spi_image, spi_err, lsmname, outlsm or lsmname, freq0=freq0, spitol=tol)
+    addSPI.addSPI(spi_image, spi_err, lsmname=lsmname, outfile=outlsm,
+                  beam=beam, freq0=freq0, spitol=tol)